configure: Check whether a compile-time integer is a power of 2
This  commit adds  some function-like  macros for  checking the
nature of a compile-time integer with respect to powers of 2,
namely the following:

  MAGICKCORE_BITS_BELOW(power_of_2)
  MAGICKCORE_IS_NOT_ALIGNED(n,power_of_2)
  MAGICKCORE_IS_NOT_POWER_OF_2(n)

These have been defined in the following header:

  MagickCore/magick-config.h

Here are the reasons for using that file:

  * It's already included by many files.

  * It  deals  with  configuration   values,  which  are  often
    the  very compile-time  integers that  we're interested  in
    testing.

These new macros have been put to use to check a couple values.
One of  these checks  involves making making  some calculations
with the value of `sizeof(void *)'; of course, `sizeof(void *)'
is  not directly  usable by  the  C Preprocessor,  and so  this
commit also uses a configure-time  test for  the value  of that
expression (provided by the last commit).

It should  be noted that the  C11 version of the  C Programming
Language provides  `static_assert', which  can be used  to make
compile-time assertions  with values such as  `sizeof(void *)'.
Also, there  are a number of  clever macro hacks to  simulate a
similar  facility  for  older  versions of  the  C  Programming
Language.  Nevertheless,   this  commit   elects  to   use  the
configure-time solution,  because it should be  the most widely
supported and understood.