coders/pcx: Remove unnecessary test
The following left-shift is made:

  (size_t)1 << (pcx_info.bits_per_pixel*pcx_info.planes)

Before that occurs, the code is dutifully checking to ensure that
the right-hand-side  of the shift  will not  be >= the  number of
bits in the representation of the  value on the left-hand side of
the shift operator (via the hard-coded `64', which should really
not be hard-coded):

  if ((pcx_info.bits_per_pixel*pcx_info.planes) >= 64)
    ThrowPCXException(CorruptImageError,"ImproperImageHeader");

However, this check has already been made implicitly by checking
the values of the operands of the multiplication:

  if ((image->columns == 0) || (image->rows == 0) ||
      ((pcx_info.bits_per_pixel != 1) && (pcx_info.bits_per_pixel != 2) &&
       (pcx_info.bits_per_pixel != 4) && (pcx_info.bits_per_pixel != 8)))
    ThrowPCXException(CorruptImageError,"ImproperImageHeader");

  ...

  if (pcx_info.planes > 6)
    ThrowPCXException(CorruptImageError,"ImproperImageHeader");

That is to say, by the time of the shift, this multiplication:

  pcx_info.bits_per_pixel*pcx_info.planes

is known to be at most 48:

  8 * 6

Therefore, the test can be omitted.

Also, 2 other  tests for `bits_per_pixel >= 8'  can be simplified
to just `bits_per_pixel == 8'.