configure: Be more careful in defining macro `_FILE_OFFSET_BITS'
The header `libraw.h' (up to at least version 0.19.5) just takes
matters into its own hands:

  #ifdef __linux__
  #define _FILE_OFFSET_BITS 64
  #endif

Hitherto,   before   including   the  header   `libraw.h',   the
ImageMagick code  would define the macro  `_FILE_OFFSET_BITS' as
follows:

  #define _FILE_OFFSET_BITS MAGICKCORE__FILE_OFFSET_BITS

Because the preprocessing-tokens do  not match, this would cause
the compiler to issue a warning that the `_FILE_OFFSET_BITS' was
redefined.

With this commit, `_FILE_OFFSET_BITS'  is now inspected and also
"sanitized" if possible to be defined as `64' or `32' directly.